restructuring (breaking down) modules in engage/twilio
